- [WIP] TYPO3 13 Support - [FEATURE] All TypoScript settings are now editable via constant editor - [CHANGE] We look up now all fe_users on registration to avoid conflicts with possible felogin procedures

+++ Classes/Controller/RegistrationController.php
@@ -67,7 +67,7 @@ class RegistrationController extends BaseActionController {
     public function registerAction(User $newUser) {        
         try {
             // Register with DOI process
-            $newUser = $this->registrationService->register($newUser, current(GeneralUtility::intExplode(',', $this->configurationManager->getContentObject()->data['pages'], true)), $this->settings);
+            $newUser = $this->registrationService->register($newUser, (int)current(GeneralUtility::intExplode(',', $this->configurationManager->getContentObject()->data['pages'], true)), $this->settings);
             $this->view->assign('mailResponse', $this->registrationService->getMailResponseForDOI());
         }catch(\Exception $e) {
             $this->view->assign('error', $e->getMessage());
